try to speed build by separating placebo tests
