+++ .travis.yml
@@ -10,6 +10,7 @@ install:
 env:
   - TESTCASE=tests/test_handler.py
   - TESTCASE=tests/tests_middleware.py
+  - TESTCASE=tests/tests_placebo.py
   - TESTCASE=tests/tests.py
 script: 
   - nosetests $TESTCASE --with-coverage --cover-package=zappa
